requirements.txt
Django==4.2.9
feedparser==6.0.11
djangorestframework==3.15.2
django-environ==0.11.2
openai==0.28
psycopg==3.2.3
drf-spectacular==0.27.2
celery==4.4.2
redis==5.2.0
websocket==0.2.1
channels==4.1.0
